feat!: check actually typed char
Neovim version 0.10.0 (nightly by the time of writing this) recently
introduce a new parameter being passed to the `vim.on_key` function
which is the "key before mapping". Although for most "insert" modes not
that relevant this comes in handy for supporting visual-mode,
select-mode and more in upcoming releases. Existing use cases should not
be affected, but this raises the minimum required Neovim version and
therefore has to be considered a breaking change.