feat: add support for operator and visual mode
To do this we have to enhance the escape function by two parameters
being the last cursor position and the last value of b:changedtick. This
allows to check for undoing certain modifications and resetting the
cursor which is needed for operator and visual mode. The default
implementation should cover most cases but there is no guarantee that
you don't need any adoptions for another combination.